# api/openapi.yaml
openapi: 3.0.3
info:
  title: Hotel Reviews Microservice API
  description: |
    A comprehensive microservice for managing hotel reviews, ratings, and related data.
    
    ## Features
    - Hotel and provider management
    - Review creation, retrieval, and management
    - Advanced search and filtering capabilities
    - File upload and batch processing
    - User authentication and authorization
    - Real-time metrics and monitoring
    
    ## Authentication
    This API uses JWT Bearer token authentication. Obtain a token by calling the `/auth/login` endpoint.
    
    ## Rate Limiting
    API requests are rate-limited. Check the `X-RateLimit-*` headers in responses for current limits.
    
    ## Versioning
    The API uses URL path versioning (e.g., `/api/v1/`). Current version is v1.
  version: 1.0.0
  contact:
    name: Hotel Reviews API Support
    email: support@hotelreviews.com
    url: https://docs.hotelreviews.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://hotelreviews.com/terms
servers:
  - url: https://api.hotelreviews.com/api/v1
    description: Production server
  - url: https://staging-api.hotelreviews.com/api/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Development server
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email verification
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
            examples:
              user_registration:
                summary: Standard user registration
                value:
                  username: "johndoe"
                  email: "john.doe@example.com"
                  password: "SecurePassword123!"
                  first_name: "John"
                  last_name: "Doe"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                registration_success:
                  summary: Successful registration
                  value:
                    success: true
                    data:
                      user:
                        id: "123e4567-e89b-12d3-a456-426614174000"
                        username: "johndoe"
                        email: "john.doe@example.com"
                        first_name: "John"
                        last_name: "Doe"
                        is_verified: false
                      access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      expires_in: 3600
                    timestamp: "2024-01-15T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and obtain access tokens
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              login_with_email:
                summary: Login with email
                value:
                  email: "john.doe@example.com"
                  password: "SecurePassword123!"
              login_with_username:
                summary: Login with username
                value:
                  username: "johndoe"
                  password: "SecurePassword123!"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Obtain a new access token using a refresh token
      operationId: refreshToken
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  description: Valid refresh token
              required:
                - refresh_token
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate current access and refresh tokens
      operationId: logoutUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Review Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from the authentication endpoints
  schemas:
...
    UserRegistrationRequest:
      type: object
      required:
        - username
        - email
        - password
        - first_name
        - last_name
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^[a-zA-Z0-9_]+$'
          description: Unique username
        email:
          type: string
          format: email
          maxLength: 255
          description: User email address
        password:
          type: string
          minLength: 8
          maxLength: 128
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]'
          description: Password with at least 8 characters, including uppercase, lowercase, number, and special character
        first_name:
          type: string
          maxLength: 100
          description: User's first name
        last_name:
          type: string
          maxLength: 100
          description: User's last name
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User email address
        username:
          type: string
          description: Username
        password:
          type: string
          description: User password
      required:
        - password
      oneOf:
        - required: [email]
        - required: [username]
    AuthResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                user:
                  $ref: '#/components/schemas/User'
                access_token:
                  type: string
                  description: JWT access token
                refresh_token:
                  type: string
                  description: JWT refresh token
                expires_in:
                  type: integer
                  description: Token expiration time in seconds
    TokenResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                access_token:
                  type: string
                  description: New JWT access token
                expires_in:
                  type: integer
                  description: Token expiration time in seconds

    # User Schemas
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User ID
        username:
          type: string
          description: Username
        email:
          type: string
          format: email
          description: User email
        first_name:
          type: string
          description: First name
        last_name:
          type: string
          description: Last name
        is_active:
          type: boolean
          description: Whether user account is active
        is_verified:
          type: boolean
          description: Whether user email is verified
        last_login_at:
          type: string
          format: date-time
          nullable: true
          description: Last login timestamp
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp

    # Review Schemas
...
    BaseResponse:
      type: object
      required:
        - success
        - timestamp
      properties:
        success:
          type: boolean
          description: Whether the request was successful
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
        trace_id:
          type: string
          nullable: true
          description: Request trace ID for debugging
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              nullable: true
              description: Response data
            message:
              type: string
              description: Success message
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - error
            - error_code
          properties:
            success:
              type: boolean
              enum: [false]
            error:
              type: string
              description: Error message
            error_code:
              type: string
              description: Machine-readable error code
            details:
              type: object
              nullable: true
              description: Additional error details
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            details:
              type: object
              properties:
                validation_errors:
                  type: array
                  items:
                    type: object
                    properties:
                      field:
                        type: string
                        description: Field that failed validation
                      message:
                        type: string
                        description: Validation error message
                      code:
                        type: string
                        description: Validation error code

    # Utility Schemas
...
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_parameter:
              summary: Invalid parameter
              value:
                success: false
                error: "Invalid parameter value"
                error_code: "INVALID_PARAMETER"
                timestamp: "2024-01-15T10:30:00Z"
                trace_id: "abc123def456"
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing_token:
              summary: Missing authentication token
              value:
                success: false
                error: "Authentication required"
                error_code: "UNAUTHORIZED"
                timestamp: "2024-01-15T10:30:00Z"
                trace_id: "abc123def456"
            invalid_token:
              summary: Invalid authentication token
              value:
                success: false
                error: "Invalid or expired token"
                error_code: "INVALID_TOKEN"
                timestamp: "2024-01-15T10:30:00Z"
                trace_id: "abc123def456"
...
    Conflict:
      description: Conflict - resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            duplicate_resource:
              summary: Resource already exists
              value:
                success: false
                error: "A resource with this identifier already exists"
                error_code: "CONFLICT"
                timestamp: "2024-01-15T10:30:00Z"
                trace_id: "abc123def456"
    ValidationError:
      description: Validation error - invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          examples:
            validation_failure:
              summary: Validation errors
              value:
                success: false
                error: "Request validation failed"
                error_code: "VALIDATION_ERROR"
                details:
                  validation_errors:
                    - field: "email"
                      message: "Invalid email format"
                      code: "INVALID_FORMAT"
                    - field: "rating"
                      message: "Rating must be between 1 and 5"
                      code: "OUT_OF_RANGE"
                timestamp: "2024-01-15T10:30:00Z"
                trace_id: "abc123def456"
    TooManyRequests:
      description: Too many requests - rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Request limit per time window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Time when rate limit resets
          schema:
            type: integer
        Retry-After:
          description: Number of seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            rate_limit_exceeded:
              summary: Rate limit exceeded
              value:
                success: false
                error: "Rate limit exceeded"
                error_code: "RATE_LIMIT_EXCEEDED"
                timestamp: "2024-01-15T10:30:00Z"
                trace_id: "abc123def456"
...